requirements.txt
amqp==2.5.2
asgiref==3.2.3
bcrypt==3.1.7
beautifulsoup4==4.8.1
billiard==3.6.1.0
bs4==0.0.1
celery==4.4.0
cffi==1.13.2
cryptography==2.8
Django==3.0.1
django-appconf==1.0.3
django-macaddress==1.6.0
django-select2==7.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
getmac==0.8.2
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
netaddr==0.7.19
netifaces==0.10.9
paramiko==2.7.1
psycopg2-binary==2.8.4
pycparser==2.19
PyNaCl==1.3.0
python-decouple==3.3
python-memcached==1.59
python-nmap==0.6.1
pytz==2019.3
redis==3.3.11
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
